requirements.txt
amqp==5.1.1
anyio==3.6.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.85.2
flower==1.2.0
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
humanize==4.4.0
idna==3.4
kombu==5.2.4
prometheus-client==0.15.0
prompt-toolkit==3.0.31
pydantic==1.10.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tornado==6.2
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.4
